+++ plugins/time-saver/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Adding customizable headers for TimeSaverPage
 - remove `Time Saved - %` prefix from All Charts Bar
+- Optional configuration to show Stats Table summaries in days
 
 ## 2.0.1